wic: Introduce the `WicExecError` exception class
Whenever errors are detected by the `fsck` utility, it will return a non
zero error code even when the errors have been self repaired, which breaks
the image generation process.

This commit implements a `wic` exception used only when executing commands,
which carries information about the command used, its return code and output,
in order to filter out errors that are not critical to image generation.